refactored directory, proxy package added
