--- src/topic_0/FastThing.java
@@ -1,5 +1,0 @@
-package topic_0;
-
-public class FastThing extends Thing {
-	public FastThing() {}
-}
