--- src/topic_0/Proxy.java
@@ -1,20 +1,0 @@
-package topic_0;
-
-import java.util.Date;
-
-public class Proxy {
-
-	SlowThing slowThing;
-
-	public Proxy() {
-		System.out.println("Creating proxy at " + new Date());
-	}
-
-	public void sayHello() {
-		if (slowThing == null) {
-			slowThing = new SlowThing();
-		}
-		slowThing.sayHello();
-	}
-
-}
