--- src/topic_0/SlowThing.java
@@ -1,12 +1,0 @@
-package topic_0;
-
-public class SlowThing extends Thing {
-	
-	public SlowThing() {
-		try {
-			Thread.sleep(5000);
-		} catch (InterruptedException e) {
-			e.printStackTrace();
-		}
-	}
-}
