--- src/topic_0/Thing.java
@@ -1,10 +1,0 @@
-package topic_0;
-
-import java.util.Date;
-
-public abstract class Thing {
-	public void sayHello() {
-		System.out.println(this.getClass().getSimpleName() + " says howdy at " + new Date());
-	}
-
-}
