+++ topic_0/proxy/FastThing.java
@@ -1,0 +1,5 @@
+package proxy;
+
+public class FastThing extends Thing {
+	public FastThing() {}
+}
