+++ topic_0/proxy/Proxy.java
@@ -1,0 +1,20 @@
+package proxy;
+
+import java.util.Date;
+
+public class Proxy {
+
+	SlowThing slowThing;
+
+	public Proxy() {
+		System.out.println("Creating proxy at " + new Date());
+	}
+
+	public void sayHello() {
+		if (slowThing == null) {
+			slowThing = new SlowThing();
+		}
+		slowThing.sayHello();
+	}
+
+}
