+++ topic_0/proxy/SlowThing.java
@@ -1,0 +1,12 @@
+package proxy;
+
+public class SlowThing extends Thing {
+	
+	public SlowThing() {
+		try {
+			Thread.sleep(5000);
+		} catch (InterruptedException e) {
+			e.printStackTrace();
+		}
+	}
+}
