+++ topic_0/proxy/Thing.java
@@ -1,0 +1,10 @@
+package proxy;
+
+import java.util.Date;
+
+public abstract class Thing {
+	public void sayHello() {
+		System.out.println(this.getClass().getSimpleName() + " says howdy at " + new Date());
+	}
+
+}
